# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/limit/slidewindowsreject": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "limit"
                ],
                "summary": "滑动窗口计数拒流，每秒超过指定次数会拒流掉",
                "responses": {
                    "200": {
                        "description": "成功会返回ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "失败返回reject"
                    }
                }
            }
        },
...
    }
}